Enforced SRP and refactored the code
Previous commit was causing bugs when more than one token was on the same
canvas.

This commit refactors the code responsible to check the token position
and the tile occlusion. The code has been broken down into smaller
functions and classes to make it easier to understand and maintain.

The new classes are:
- OcclusionManager: responsible for checking and updating the occlusion of
 the tiles on the canvas conditional on the token position.
- TokenTileUtils: responsible for checking the position of a token
    with respect to the tiles on the canvas.

The occlusion mode of a tile is now fievaluatedred only when a token is
moved or updated (hooking on `updateToken`) or selected (hooking on
`controlToken`).

Also, the selected token and tiles are now evaluated only when
respectibely a token is selected or unselected (hooking on `controlToken`),
and when a tile settings are changed (hooking on `updateTile`). This should
reduce the number of times the occlusion is evaluated and improve the
performance of the module.